refactor: extraer selector_helpers.h y ajustar forward-decls del refactor lowering
- selector_helpers.h: SCRATCH_A/B/C, JIT_PROC_REG, NATIVE_ARG0/1/2
- emit_field_addr con offset uint64_t y forward-decls sin static
- tests/gc/test_monitor_cas.cpp ajustado

+++ src/vex/lowering_class.cpp
@@ -1403,7 +1403,7 @@ ir::IrValueId emit_field_addr(ir::IrFunction *fn, ir::IrBlockId block,
     ir::IrValueId emit_field_addr(ir::IrFunction *fn,
                                          ir::IrBlockId   block,
                                          ir::IrValueId   base,
-                                         uint32_t        offset,
+                                         uint64_t        offset,
                                          uint32_t        line) {
         if (offset == 0) {
             // El frontend marca el resultado como host_ptr para que LOAD/
